+++ src/dialogs/mainwindow.cpp
@@ -8,11 +8,10 @@
 #include "customwidgets/notesdock.h"
 #include "customwidgets/statusbar.h"
 #include "dialogs/aboutdialog.h"
-#include "dialogs/connectiondialog.h"
 #include "dialogs/diagnosticdialog.h"
 #include "dialogs/importmbcdialog.h"
-#include "dialogs/logdialog.h"
 #include "dialogs/registerdialog.h"
+#include "dialogs/settingsdialog.h"
 #include "dialogs/ui_mainwindow.h"
 #include "graphview/graphview.h"
 #include "importexport/datafilehandler.h"
@@ -52,8 +51,6 @@ MainWindow::MainWindow(QStringList cmdArguments, GuiModel* pGuiModel,
 
     QApplication::setStyle(QStyleFactory::create("Fusion"));
 
-    _pConnectionDialog = new ConnectionDialog(_pSettingsModel, this);
-    _pLogDialog = new LogDialog(_pSettingsModel, this);
     _pDiagnosticDialog = new DiagnosticDialog(_pDiagnosticModel, this);
 
     _pNotesDock = new NotesDock(_pNoteModel, _pGuiModel, this);
@@ -100,8 +97,7 @@ MainWindow::MainWindow(QStringList cmdArguments, GuiModel* pGuiModel,
     connect(_pUi->actionHighlightSamplePoints, &QAction::toggled, _pGuiModel, &GuiModel::setHighlightSamples);
     connect(_pUi->actionClearData, &QAction::triggered, this, &MainWindow::clearData);
     connect(_pUi->actionToggleMarkers, &QAction::triggered, this, &MainWindow::toggleMarkersState);
-    connect(_pUi->actionConnectionSettings, &QAction::triggered, this, &MainWindow::showConnectionDialog);
-    connect(_pUi->actionLogSettings, &QAction::triggered, this, &MainWindow::showLogSettingsDialog);
+    connect(_pUi->actionSettings, &QAction::triggered, this, &MainWindow::showSettingsDialog);
     connect(_pUi->actionRegisterSettings, &QAction::triggered, this, &MainWindow::handleShowRegisterDialog);
     connect(_pUi->actionAddNote, &QAction::triggered, this, &MainWindow::addNoteToGraph);
     connect(_pUi->actionZoom, &QAction::triggered, this, &MainWindow::toggleZoom); /* Only called on GUI click, not on setChecked */
@@ -228,7 +224,6 @@ MainWindow::MainWindow(QStringList cmdArguments, GuiModel* pGuiModel,
 MainWindow::~MainWindow()
 {
     delete _pGraphView;
-    delete _pConnectionDialog;
     delete _pModbusPoll;
     delete _pGraphShowHide;
     delete _pMostRecentMenu;
@@ -354,14 +349,10 @@ void MainWindow::menuShowHideGraphClicked(bool bState)
     _pGraphDataModel->setVisible(graphIdx, bState);
 }
 
-void MainWindow::showConnectionDialog()
+void MainWindow::showSettingsDialog()
 {
-    _pConnectionDialog->exec();
-}
-
-void MainWindow::showLogSettingsDialog()
-{
-    _pLogDialog->exec();
+    SettingsDialog dialog(_pSettingsModel);
+    dialog.exec();
 }
 
 void MainWindow::handleShowRegisterDialog(bool checked)
@@ -668,8 +659,7 @@ void MainWindow::updateGuiState()
     if (_pGuiModel->guiState() == GuiState::INIT)
     {
         _pUi->actionStop->setEnabled(false);
-        _pUi->actionConnectionSettings->setEnabled(true);
-        _pUi->actionLogSettings->setEnabled(true);
+        _pUi->actionSettings->setEnabled(true);
         _pUi->actionRegisterSettings->setEnabled(true);
         _pUi->actionStart->setEnabled(true);
         _pUi->actionOpenDataFile->setEnabled(true);
@@ -689,8 +679,7 @@ void MainWindow::updateGuiState()
     {
         // Communication active
         _pUi->actionStop->setEnabled(true);
-        _pUi->actionConnectionSettings->setEnabled(false);
-        _pUi->actionLogSettings->setEnabled(false);
+        _pUi->actionSettings->setEnabled(false);
         _pUi->actionRegisterSettings->setEnabled(false);
         _pUi->actionStart->setEnabled(false);
         _pUi->actionOpenDataFile->setEnabled(false);
@@ -707,8 +696,7 @@ void MainWindow::updateGuiState()
     {
         // Communication not active
         _pUi->actionStop->setEnabled(false);
-        _pUi->actionConnectionSettings->setEnabled(true);
-        _pUi->actionLogSettings->setEnabled(true);
+        _pUi->actionSettings->setEnabled(true);
         _pUi->actionRegisterSettings->setEnabled(true);
         _pUi->actionStart->setEnabled(true);
         _pUi->actionOpenDataFile->setEnabled(true);
@@ -736,8 +724,7 @@ void MainWindow::updateGuiState()
     {
         // Communication not active
         _pUi->actionStop->setEnabled(false);
-        _pUi->actionConnectionSettings->setEnabled(true);
-        _pUi->actionLogSettings->setEnabled(true);
+        _pUi->actionSettings->setEnabled(true);
         _pUi->actionRegisterSettings->setEnabled(true);
         _pUi->actionStart->setEnabled(true);
         _pUi->actionOpenDataFile->setEnabled(true);